+++ dataland-frontend/src/components/resources/dataTable/MultiLayerDataTableCell.vue
@@ -1,39 +1,24 @@
 <template>
-  <template v-if="content.displayComponent == MLDTDisplayComponents.StringDisplayComponent">
-    <span v-if="content.displayValue">{{ content.displayValue }}</span>
-  </template>
-
-  <template v-else-if="isDisplayValueOfType(MLDTDisplayComponents.DocumentLinkDisplayComponent, content)">
-    <DocumentLink
-      :label="content.displayValue.label"
-      :download-name="content.displayValue.reference.name"
-      :reference="content.displayValue.reference.reference"
-      show-icon
-    />
-  </template>
-
-  <template v-else-if="isDisplayValueOfType(MLDTDisplayComponents.ModalLinkDisplayComponent, content)">
-    <a @click="dialog.open(content.displayValue.modalComponent, content.displayValue.modalOptions)" class="link"
-      >{{ content.displayValue.label }}
-      <em class="pl-2 material-icons" aria-hidden="true" title=""> dataset </em>
-    </a>
-  </template>
+  <component :is="content.displayComponent" :content="content" />
 </template>
 
-<script generic="T extends MLDTDisplayComponents" setup lang="ts">
-import {
-  isDisplayValueOfType,
-  MLDTDisplayComponents,
-  type MLDTDisplayValue,
-} from "@/components/resources/dataTable/MultiLayerDataTableConfiguration";
-import DocumentLink from "@/components/resources/frameworkDataSearch/DocumentLink.vue";
-import { useDialog } from "primevue/usedialog";
+<script lang="ts">
+import { type AvailableDisplayValues } from "@/components/resources/dataTable/MultiLayerDataTableConfiguration";
+import DocumentLinkDisplayComponent from "@/components/resources/dataTable/cells/DocumentLinkDisplayComponent.vue";
+import ModalLinkDisplayComponent from "@/components/resources/dataTable/cells/ModalLinkDisplayComponent.vue";
+import StringDisplayComponent from "@/components/resources/dataTable/cells/StringDisplayComponent.vue";
+import { defineComponent } from "vue";
 
-const dialog = useDialog();
-
-defineProps<{
-  content: MLDTDisplayValue<T>;
-}>();
+export default defineComponent({
+  name: "MultiLayerDataTableCell",
+  components: { DocumentLinkDisplayComponent, ModalLinkDisplayComponent, StringDisplayComponent },
+  props: {
+    content: {
+      type: Object as () => AvailableDisplayValues,
+      required: true,
+    },
+  },
+});
 </script>
 
 <style scoped></style>
